[hotfix][yarn] Move getDynamicProperties from yarn utils to BootstrapTools
The getDynamicProperties could be reused by KubernetesResourceManager to generate taskmanager start command.
