+++ flink-yarn/src/main/java/org/apache/flink/yarn/Utils.java
@@ -19,11 +19,9 @@
 package org.apache.flink.yarn;
 
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.configuration.ConfigOptions;
 import org.apache.flink.configuration.ResourceManagerOptions;
 import org.apache.flink.runtime.clusterframework.BootstrapTools;
 import org.apache.flink.runtime.clusterframework.ContaineredTaskManagerParameters;
-import org.apache.flink.runtime.entrypoint.parser.CommandLineOptions;
 import org.apache.flink.runtime.util.HadoopUtils;
 import org.apache.flink.util.StringUtils;
 
@@ -64,7 +62,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Stream;
 
 import static org.apache.flink.yarn.YarnConfigKeys.ENV_FLINK_CLASSPATH;
 
@@ -635,34 +632,6 @@ public final class Utils {
 		if (!condition) {
 			throw new RuntimeException(String.format(message, values));
 		}
-	}
-
-	/**
-	 * Get dynamic properties based on two Flink configuration. If base config does not contain and target config
-	 * contains the key or the value is different, it should be added to results. Otherwise, if the base config contains
-	 * and target config does not contain the key, it will be ignored.
-	 * @param baseConfig The base configuration.
-	 * @param targetConfig The target configuration.
-	 * @return Dynamic properties as string, separated by space.
-	 */
-	static String getDynamicProperties(
-		org.apache.flink.configuration.Configuration baseConfig,
-		org.apache.flink.configuration.Configuration targetConfig) {
-
-		String[] newAddedConfigs = targetConfig.keySet().stream().flatMap(
-			(String key) -> {
-				final String baseValue = baseConfig.getString(ConfigOptions.key(key).stringType().noDefaultValue());
-				final String targetValue = targetConfig.getString(ConfigOptions.key(key).stringType().noDefaultValue());
-
-				if (!baseConfig.keySet().contains(key) || !baseValue.equals(targetValue)) {
-					return Stream.of("-" + CommandLineOptions.DYNAMIC_PROPERTY_OPTION.getOpt() + key +
-						CommandLineOptions.DYNAMIC_PROPERTY_OPTION.getValueSeparator() + targetValue);
-				} else {
-					return Stream.empty();
-				}
-			})
-			.toArray(String[]::new);
-		return String.join(" ", newAddedConfigs);
 	}
 
 }
